updating notes and restructuring modules
