+++ api/database.js
@@ -85,98 +85,4 @@ async function refreshLiveData(sessionID, sessionUsersArray, category, tableList
     }
 }
 
-/**
- * 
- * @param {string} username - username of user who will be added to the session
- * @param {string} sessionID  - the session which should they be added to
- * @param {array} DB_SESSIONS - the array of session data
- * @returns 
- */
-async function joinSession(username, sessionID, sessions = []) {
-
-    // Access server and push username to that open session
-    // Without access to server's live data, access and push new data to JSON
-    for (let { session_id, active_users_array, end_time } of sessions) {
-        if (sessionID === session_id && end_time === '') {
-            
-            // check if user is already added to server
-            const userInSession = active_users_array.find(user => user.name === username)
-
-            if (!userInSession) {
-                // Add user to the active users array
-                active_users_array.push({name: username, vote: null})
-            }
-
-            if (userInSession) {
-                // Remove the user's vote from the array
-                userInSession['vote'] = null
-            }
-
-            // Refresh the database with the updated data
-            await refreshDatabase(sessions, null, null)
-
-            // User added to session (if not already) and database is updated
-            return true
-        }
-    }
-    
-    // Session is closed or does not exist
-    return false
-}
-
-/** Create a session id and verify that it doesn't already exist
- * 
- * @param {string} sessionCategory - the category (movie, game, food) to send to randomSessionID()
- * @param {array} DB_SESSIONS - array of sessions to check for duplicates
- * @returns the new session id
- */
-function createSessionID(sessionCategory, DB_SESSIONS) {
-    let newSessionID
-    for (let loopLimit = 0; loopLimit < 10; loopLimit++) {
-        
-        // Create a new session ID
-        newSessionID = randomSessionID(sessionCategory)
-
-        // If it already exists, make a new one
-        const existingID = DB_SESSIONS.some(session => session.session_id === newSessionID)
-        if (!existingID) {return newSessionID}
-    }
-    console.log('Something went wrong while creating a session ID.')
-    return null
-}
-
-/** Create a random session id using the category
- * 
- * @param {string} sessionCategory - the category (movie, game, food) for which a session will be created
- * @returns the created session id
- */
-function randomSessionID(sessionCategory) {
-    let sessionString = []
-    let digitArray1
-
-    switch (true) {
-        case sessionCategory === 'food':
-            digitArray1 = ['B', 'C', 'D', 'F', 'G', 'H', 'J']
-            break
-        case sessionCategory === 'movie':
-            digitArray1 = ['K', 'L', 'M', 'N', 'P', 'Q']
-            break
-        case sessionCategory === 'game':
-            digitArray1 = ['R', 'S', 'T', 'V', 'W', 'X', 'Z']
-            break
-    }
-    let digitArray2 = 'BCDFGHJKLMNPQRSTVWXZ'.split('')
-    let digitArray3 = '23456789'.split('')
-
-    sessionString.push(classes.randomDigit(digitArray1))
-    sessionString.push(classes.randomDigit(digitArray3))
-    sessionString.push(classes.randomDigit(digitArray2))
-    sessionString.push(classes.randomDigit(digitArray3))
-    sessionString.push(classes.randomDigit(digitArray2))
-    sessionString.push(classes.randomDigit(digitArray3))
-
-    sessionString = sessionString.join('')
-    return sessionString
-}
-
-module.exports = { loadDatabase, refreshDatabase, refreshLiveData, joinSession, createSessionID }
+module.exports = { loadDatabase, refreshDatabase, refreshLiveData }
