+++ api/login.js
@@ -3,56 +3,58 @@ const classes = require('./classes.js')
 
 function pageSetup (app) {
 
-    // Backend function to validate login information of existing user
+// 1.1 -- Validate current user login
     app.get('/api/validate-login/:username/:password', async (req, res) => {
+        
+// 1.1.1 -- Gather information inputted from login page
         const checkUsername = req.params.username
         const checkPassword = req.params.password
         
-        // Get users database
+// 1.1.2 ---- Compare against database of current users
+// 1.1.2.1 -- Request persistent database data
         const data = await database.loadDatabase()
         const DB_USERS = data.users
 
-        // Initialize the check figures
+// 1.1.2.2 -- Compare given username and recorded username
         let goodUsername = false
         let goodPassword = false
 
-        // Compare given username and recorded usernames
         for ( let { username, salt, password_hash } of DB_USERS) {
             if (username === checkUsername) {
-                // When one matches
+
                 goodUsername = true
 
-                // Compare hash of given password and recorded salt with recorded password_hash
+// 1.1.2.3 ---- Compare hash of given password and recorded salt with recorded password_hash
                 const checkHash = classes.hashPassword(checkPassword, salt)
                 password_hash === checkHash ? goodPassword = true : goodPassword = false
             }
         }
 
-        // Return the results
         const validLogin = {
             goodUsername: goodUsername,
             goodPassword: goodPassword
         }
+
         res.json(validLogin)
+
     })
 
-
-    // Backend function to create login information for new user
+// 1.2 -- Create new user
     app.get('/api/create-login/:username/:password/:confirmation', async (req, res) => {
+
+// 1.2.1 -- Gather information inputted from login page
         const checkUsername = req.params.username
         const checkPassword = req.params.password
         const checkConfirmation = req.params.confirmation
-
-        // Initialize the check figures
-        let goodUsername = true
-        let goodPassword = false
-        let goodConfirmation = false
         
-        // Get users database
+// 1.2.2 ---- Compare against database of current users
+// 1.2.2.1 -- Request persistent database data (dummy_data.json)
         const data = await database.loadDatabase()
         const DB_USERS = data.users
 
-        // Username must be unique
+// 1.2.2.2 -- Compare given username and all recorded usernames
+        let goodUsername = true
+
         for (let { username } of DB_USERS) {
             if (checkUsername === username) {
                 goodUsername = false
@@ -60,20 +62,25 @@ function pageSetup (app) {
             } 
         }
 
-        // Password must be 8_ characters with 1+ letter(s) and 1+ number(s)
+// 1.2.3 -- Ensure given password complies with password requirements
+        let goodPassword = false
+        let goodConfirmation = false
+
+// 1.2.3.1 -- Password must be 8_ characters with 1+ letter(s) and 1+ number(s)
         const passwordRegex = /^(?=.*[a-zA-Z])(?=.*\d).{8,}$/
         goodPassword = passwordRegex.test( checkPassword )
 
-        // Password confirmation must match the given password
+// 1.2.3.2 -- Password confirmation must match the given password
         goodConfirmation = ( checkPassword === checkConfirmation )
 
+// 1.2.5 -- If info is good, create new user with username and password, proceed to 2
         if (goodUsername && goodPassword && goodConfirmation) {
 
-            // Create new user and add to database
+// 1.2.5.1 -- Create new user
             const createUser = new classes.User(checkUsername, checkPassword)
             DB_USERS.push(createUser)
 
-            // Save the updated JSON data to the database
+// 1.2.5.2 -- Send to persistent database (dummy_data.json)
             database.refreshDatabase(null, DB_USERS, null)
         }
 
@@ -84,6 +91,7 @@ function pageSetup (app) {
         }
         
         res.json(createLogin)
+
     })
 
 }
